% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropmodel.R
\name{dropmodel}
\alias{dropmodel}
\title{Fits the drop-model or missingness-model using logistic regression for estimating and computing inverse probability censoring weights.}
\usage{
dropmodel(formula, data, cluster.var, unit.var, time.var)
}
\arguments{
\item{formula}{a formula expression for the drop-out model. The outcome should be the missingness indicator (binary).}

\item{data}{an optional data frame containing the variables provided in
\code{formula}, \code{cluster.var}, \code{unit.var} and \code{time.var}.}

\item{cluster.var}{a vector that identifies the clusters.}

\item{unit.var}{a vector that identifies the unit within a cluster.}

\item{time.var}{a vector that identifies the repeated observation of a unit.}
}
\value{
Returns a list with components:
\item{call}{the matched call.}
\item{coefficients}{the estimated regression parameters, standard errors, z-values, and  p-values of the logistic regression model, fitted via \code{glm}.}
\item{outdata}{the original data set used in the function plus computed IPCW (variable name "ipcw") as a new variable.}
\item{Slist}{the cluster-level score functions that will be used to compute the robust variance in the final model.}
}
\description{
\code{dropmodel} fits a logitistic regression model via \code{\link[stats]{glm}} to predict drop-out or missingness
to estimate and compute inverse probability censoring weights (IPCWs).
The output contains
1. Coeffficient estimates and inference,
2. A data set that contains all
original variables and computed IPCWs (variable name \code{ipcw}), and
3. The cluster-level score functions that will  be used to compute the robust variance in the final model.
Users can proceed to \code{\link{ipccwGEE}} to fit a marginal model via inverse probability censoring cluster-weighted generalized estimating equations.
}
\details{
The \code{data} must be provided in case level or equivalently in `long' format.
}
\examples{
data(dental)
outdrop <- dropmodel(toothstat ~ prevmaxcal5mm + basenumteeth, data = dental,
cluster.var = subject, unit.var = tooth, time.var = visit)
outdat <- outdrop$outdata
Slist <- outdrop$Slist
}
\author{
Aya Mitani
}
