% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipccwGEE.R
\name{ipccwGEE}
\alias{ipccwGEE}
\title{Inverse probability censoring cluster weighted GEE for clustered longitudinal data with informative cluster size and informative drop-out.}
\usage{
ipccwGEE(formula, data, cluster.var, unit.var, ipcw, Slist,
  show.iter = FALSE)
}
\arguments{
\item{formula}{a formula expression as for the regression model.}

\item{data}{an optional data frame containing the variables provided in
\code{formula}, \code{cluster.var}, \code{unit.var} and \code{ipcw}.}

\item{cluster.var}{a vector that identifies the clusters.}

\item{unit.var}{a vector that identifies the unit within a cluster.}

\item{ipcw}{a vector that contains the inverse probability censoring weights for each observation, computing from \code{dropmodel}.}

\item{Slist}{a list that contains the cluster-level score functions from the drop-out model fitted using \code{dropmodel}.}

\item{show.iter}{logical. If \code{'TRUE'}, displays the iteration number during the estimation process.}
}
\value{
Returns an object of the class \code{"ipccwgee"}. This has components:
\item{call}{the matched call.}
\item{coefficients}{the estimated regression parameter vector of the marginal model.}
\item{coef.names}{the variable name of the coefficients.}
\item{robust.variance}{the estimated "robust" covariance matrix.}
\item{robust.se}{the estimated "robust" standard errors.}
\item{wald.chisq}{the Wald Chi-square test statistic for coefficient estimates.}
\item{p.value}{the p-value based on a Wald Chi-square test statistic that no covariates are statistically significant.}
\item{niter}{the number of iterations the model took to converge.}
}
\description{
Solves the cluster-weighted generalized estimating equations
with inverse probability censoring weights for correlated binary
responses in clustered longitudinal data with informative cluster size and informative drop-out
assuming a logit model for the marginal probabilities.
}
\details{
The \code{data} must be provided in case level or equivalently in `long' format.
}
\examples{
data(dental)
## First fit the drop model or missingness model.
outdrop <- dropmodel(toothstat ~ prevmaxcal5mm + basenumteeth, data = dental,
cluster.var = subject, unit.var = tooth, time.var = visit)
## Save the outputs.
outdat <- outdrop$outdata
Slist <- outdrop$Slist
## Restrict data set to include visits where the unit (tooth) was observed.
subdental <- subset(outdat, outdat$toothstat == 1)
## Fit marginal model with cluster-weights and IPCWs.
ipccwGEEout <- ipccwGEE(maxcal5mm ~ baseage + edu, data = subdental, cluster = subject, unit = tooth,
ipcw = ipcw, Slist = Slist, show.iter = TRUE)
summary(ipccwGEEout)
}
\author{
Aya Mitani
}
